Introducing authenticate() method on the realm, and moving code to both old and new store
